Eliminate browser-shims from guessFeat tests
