+++ root/static/scripts/tests/guessFeat.js
@@ -3,8 +3,6 @@
 // Licensed under the GPL version 2, or (at your option) any later version:
 // http://www.gnu.org/licenses/gpl-2.0.txt
 
-require('./browser-shims');
-
 const _ = require('lodash');
 const test = require('tape');
